feat: add post delete (dao,service...)

+++ src/main/java/com/example/post/dao/PostDao.java
@@ -3,7 +3,7 @@ package com.example.post.dao;
 import com.example.post.model.Post;
 import com.example.post.model.UserLogin;
 
-public interface PostDao extends PostGenericDao<Post, Long> {
+public interface PostDao extends PostGenericDao<Post, Integer> {
     Post findAllPosts(String title, String author, String content);
 
     Post createAPost(String title, String author, String content);
